+++ eq-author/src/App/settings/ThemesPage.js
@@ -5,7 +5,7 @@ import { withRouter, useParams } from "react-router-dom";
 import { useMutation } from "@apollo/react-hooks";
 
 import updateQuestionnaireMutation from "graphql/updateQuestionnaire.graphql";
-import updateTheme from "graphql/updateTheme.graphql";
+
 import { colors } from "constants/theme";
 
 import VerticalTabs from "components/VerticalTabs";
@@ -18,6 +18,9 @@ import Header from "components/EditorLayout/Header";
 import ScrollPane from "components/ScrollPane";
 import { Field, Input, Label } from "components/Forms";
 import { Grid, Column } from "components/Grid";
+
+import FormType from "./FormType";
+import EqId from "./EqId";
 
 import { THEME_TITLES } from "constants/themeSettings";
 import { THEME_ERROR_MESSAGES } from "constants/validationMessages";
@@ -48,23 +51,15 @@ const PageContainer = styled.div`
   border-left: 1px solid ${colors.lightGrey};
 `;
 
-const StyledPanel = styled.div`
+const Panel = styled.div`
   max-width: 97.5%;
   padding: 1.3em;
 `;
 
-const StyledIdContainerOuter = styled.div`
+const IdContainer = styled.div`
   overflow: hidden;
   padding: 0 0 4px 4px;
-`;
-
-const StyledEqIdContainer = styled.div`
-  float: left;
-`;
-
-const StyledFormTypeContainer = styled.div`
-  margin-left: 1em;
-  float: left;
+  margin-top: 1em;
 `;
 
 const StyledInput = styled(Input)`
@@ -83,68 +78,22 @@ const HorizontalSeparator = styled.hr`
   margin: 1.5em 0;
 `;
 
-const EqIdInput = ({ eqId = "", questionnaireId, shortName }) => {
-  const [state, setState] = useState(eqId);
-  const [updateQuestionnaireTheme] = useMutation(updateTheme);
+const Heading = styled.h2`
+  font-size: 1em;
+  margin-top: 0;
+`;
 
-  const handleEQIdBlur = ({ value }, shortName) =>
-    updateQuestionnaireTheme({
-      variables: {
-        input: { questionnaireId, shortName, eqId: value.trim() },
-      },
-    });
-
-  return (
-    <StyledInput
-      value={state}
-      onChange={({ value }) => setState(value)}
-      onBlur={(e) => handleEQIdBlur({ ...e.target }, shortName)}
-      data-test={`${shortName}-eq-id-input`}
-    />
-  );
-};
-
-const FormTypeInput = ({ formType = "", questionnaireId, shortName }) => {
-  const [state, setState] = useState(formType);
-  const [updateQuestionnaireTheme] = useMutation(updateTheme);
-
-  const handleFormTypeBlur = ({ value }, shortName) => {
-    value = value.trim();
-    updateQuestionnaireTheme({
-      variables: {
-        input: { questionnaireId, shortName, formType: value },
-      },
-    });
-  };
-
-  return (
-    <StyledInput
-      value={state}
-      onChange={({ value }) => setState(value)}
-      onBlur={(e) => handleFormTypeBlur({ ...e.target }, shortName)}
-      data-test={`${shortName}-form-type-input`}
-    />
-  );
-};
-
-EqIdInput.propTypes = {
-  eqId: PropTypes.string,
-  questionnaireId: PropTypes.string,
-  shortName: PropTypes.string.isRequired,
-};
-
-FormTypeInput.propTypes = {
-  formType: PropTypes.string,
-  questionnaireId: PropTypes.string,
-  shortName: PropTypes.string.isRequired,
-};
+const Text = styled.p``;
 
 const ThemesPage = ({ questionnaire }) => {
   const { type, surveyId, id, themeSettings } = questionnaire;
+  const { themes: questionnaireThemes } = themeSettings;
+
   const [updateQuestionnaire] = useMutation(updateQuestionnaireMutation);
+  const [enableTheme] = useMutation(enableThemeMutation);
+  const [disableTheme] = useMutation(disableThemeMutation);
   const [questionnaireId, setQuestionnaireId] = useState(surveyId);
   const params = useParams();
-  const { themes: questionnaireThemes } = themeSettings;
 
   const handleBlur = ({ value }) => {
     value = value.trim();
@@ -153,17 +102,14 @@ const ThemesPage = ({ questionnaire }) => {
     });
   };
 
-  const [enableTheme] = useMutation(enableThemeMutation);
-  const [disableTheme] = useMutation(disableThemeMutation);
-
-  const themeErrorCount = themeSettings.validationErrorInfo?.totalCount ?? 0;
-
   const toggleTheme = ({ shortName, enabled }) => {
     const mutation = enabled ? disableTheme : enableTheme;
     mutation({
       variables: { input: { questionnaireId: id, shortName } },
     });
   };
+
+  const themeErrorCount = themeSettings.validationErrorInfo?.totalCount ?? 0;
 
   const groupErrorMessages = themeSettings.validationErrorInfo.errors
     .filter(({ type }) => type === "themeSettings")
@@ -187,26 +133,19 @@ const ThemesPage = ({ questionnaire }) => {
               />
               <Column gutters={false} cols={9.5}>
                 <SettingsContainer>
-                  <StyledPanel>
-                    <Field>
-                      <Label>Themes, IDs, form types and legal bases</Label>
-                    </Field>
-                    <Field>
-                      <p data-test="theme-description">
-                        The theme sets the design of the eQ for respondents. It
-                        changes the header across the survey, as well as the
-                        contact details and the legal basis on the introduction
-                        page. The COVID theme also changes the thank you page
-                        respondents see once they&apos;ve submitted the survey.
-                      </p>
-                    </Field>
-                    <Field>
-                      <p>
-                        The preview theme is applied when you view the survey
-                        using the View Survey button.
-                      </p>
-                    </Field>
-
+                  <Panel>
+                    <Heading>Themes, IDs, form types and legal bases</Heading>
+                    <Text data-test="theme-description">
+                      The theme sets the design of the eQ for respondents. It
+                      changes the header across the survey, as well as the
+                      contact details and the legal basis on the introduction
+                      page. The COVID theme also changes the thank you page
+                      respondents see once they&apos;ve submitted the survey.
+                    </Text>
+                    <Text>
+                      The preview theme is applied when you view the survey
+                      using the View Survey button.
+                    </Text>
                     <Field>
                       <Label>Survey ID</Label>
                       <Caption>
@@ -235,33 +174,18 @@ const ThemesPage = ({ questionnaire }) => {
                           onChange={() => toggleTheme({ shortName, enabled })}
                           data-test={`${shortName}-toggle`}
                         >
-                          <p />
-                          <StyledIdContainerOuter>
-                            <StyledEqIdContainer>
-                              <Field>
-                                <Label>eQ ID</Label>
-                              </Field>
-                              <EqIdInput
-                                eqId={eqId}
-                                questionnaireId={id}
-                                shortName={shortName}
-                              />
-                            </StyledEqIdContainer>
-                            <StyledFormTypeContainer>
-                              <Field>
-                                <Label>Form type</Label>
-                              </Field>
-                              <FormTypeInput
-                                formType={formType}
-                                questionnaireId={id}
-                                shortName={shortName}
-                              />
-                            </StyledFormTypeContainer>
-                          </StyledIdContainerOuter>
+                          <IdContainer>
+                            <EqId eId={eqId} qId={id} shortName={shortName} />
+                            <FormType
+                              formType={formType}
+                              qId={id}
+                              shortName={shortName}
+                            />
+                          </IdContainer>
                         </CollapsibleToggled>
                       )
                     )}
-                  </StyledPanel>
+                  </Panel>
                 </SettingsContainer>
               </Column>
             </Grid>
